% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelp.R
\docType{data}
\name{datMC}
\alias{datMC}
\title{Monte-Carlo pre-generated random integration points}
\format{This data frame contains the following columns: 
\describe{
\item{X.0.61304335481364680}{50000 pre-generated random gaussian points}
}}
\usage{
data(datMC)
}
\description{
This is a set of 50000 pre-generated random Gaussian numbers
used to be able to reproduce results across different supports
(Fortran RNGs gives different results between GNU V.4 and V.8)
}
