% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalOpenMPFortran.R
\name{evalOpenMPFortran}
\alias{evalOpenMPFortran}
\title{Evaluated computer time occording to the number of threads used to compute an arithmetic sequence. 
We used a Fortran code  combined with OpenMP}
\usage{
evalOpenMPFortran(nbrthread = 5, maxSum=100000000)
}
\arguments{
\item{nbrthread}{The mximum number ot threads to considered}

\item{maxSum}{The the maximum term of the suite}
}
\value{
a dataframe containing the computer times associated to each number of thread
}
\description{
The main objective of this function was to compare computer burden when increase the number of thread. 
In the example, we observe inthe associated plot that when the maximum term of the sequence is very large, 
we do not decrease anymore the computer time with a number of threads set greater than the number of core
}
\examples{

\dontrun{

###- with default parameters
evalOpenMPFortran()

###- with given parameters
evalOpenMPFortran(5,100000000)

}

}
\seealso{
\code{\link{testOpenMPFortran}}
}
\author{
Casimir Ledoux SOFEU <scl.ledoux@gmail.com>
}
\keyword{Fortran}
\keyword{OpenMP}
\keyword{R}
