% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrant.R
\name{funcJointSurroKendall}
\alias{funcJointSurroKendall}
\title{Integrant pour le calcul du taux de kendall pour le modele conjoint de validation d'un 
critere de substitution, en utilisant R}
\usage{
funcJointSurroKendall(u, w, up, wp, theta, gamma, alpha, eta, ui = 1)
}
\arguments{
\item{u}{f}

\item{w}{f}

\item{up}{f}

\item{wp}{f}

\item{theta}{f}

\item{gamma}{f}

\item{alpha}{f}

\item{eta}{f}

\item{ui}{f}
}
\description{
Integrant pour le calcul du taux de kendall pour le modele conjoint de validation d'un 
critere de substitution, en utilisant R
}
