% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointSurroKendall.R
\name{jointSurroKendall}
\alias{jointSurroKendall}
\title{Kendall's \eqn{\tau} estimation using Gaussian Hermite quadrature integrarion}
\usage{
jointSurroKendall(theta, gamma, alpha = 1, eta = 1, adaptative = 0,
  npg = 20, ui = 1, ui.chap.Essai = rep(0, 4),
  invBi.chol.Essai.k = matrix(rep(0, 16), nrow = 4, ncol = 4))
}
\arguments{
\item{theta}{h}

\item{gamma}{h}

\item{alpha}{h}

\item{eta}{h}

\item{adaptative}{h}

\item{npg}{h}

\item{ui}{variable binaire (1 = oui, 0 = non) qui dit si on prend en compte l'heterogeneite 
sur les risque de base ou pas}

\item{ui.chap.Essai}{h}

\item{invBi.chol.Essai.k}{h}
}
\value{
Cette fonction retourne le resultat du calcul integrale par quadratur de gauss 
hermite classique ou pseudo-adaptative
}
\description{
idem a la fonction \link{jointSurroKendall}, seulement tout le code ici est fait sur R
alors qu'avec l'autre fonction, le calcul se fait sur Fortran
}
