% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointSurrSimul.R
\name{jointSurrSimul}
\alias{jointSurrSimul}
\title{Generate survival times for two endpoints using the joint frailty surrogate model}
\usage{
jointSurrSimul(n.obs = 600, n.trial = 30, cens.adm = 549.24,
  alpha = 1.5, theta = 3.5, gamma = 2.5, zeta = 1, sigma.s = 0.7,
  sigma.t = 0.7, rsqrt = 0.8, betas = -1.25, betat = -1.25,
  frailt.base = 1, lambda.S = 1.8, nu.S = 0.0045, lambda.T = 3,
  nu.T = 0.0025, ver = 1, typeOf = 1, equi.subj.trial = 1,
  equi.subj.trt = 1, prop.subj.trial = NULL, prop.subj.trt = NULL,
  full.data = 0, random.generator = 1, random = 0, random.nb.sim = 0,
  seed = 0, nb.reject.data = 0)
}
\arguments{
\item{n.obs}{Number of considered  subjects. The default is \code{600}.}

\item{n.trial}{Number of considered  trials. The default is \code{30}.}

\item{cens.adm}{censorship time. The default is \code{549}, for about \code{40\%} of censored subjects.}

\item{alpha}{Fixed value for \eqn{\alpha}. The default is \code{1.5}.}

\item{theta}{Fixed value for \eqn{\theta}. The default is \code{3.5}.}

\item{gamma}{Fixed value for \eqn{\gamma}. The default is \code{2.5}.}

\item{zeta}{Fixed value for \eqn{\zeta}. The default is \code{1}.}

\item{sigma.s}{Fixed value for \eqn{\sigma^2_S}. The default is \code{0.7}.}

\item{sigma.t}{Fixed value for \eqn{\sigma^2_T}. The default is \code{0.7}.}

\item{rsqrt}{Desired level of correlation between \eqn{v_{S_i}} and \eqn{v_{T_i}}. \eqn{R^2_{trial}=rsqrt^2}. 
The default is \code{0.8}.}

\item{betas}{Fixed value for \eqn{\beta_S}. The default is \code{-1.25}.}

\item{betat}{Fixed value for \eqn{\beta_T}. The default is \code{-1.25}.}

\item{frailt.base}{considered the heterogeneity on the baseline risk \code{(1)} or not \code{(0)}. 
The default is \code{1}.}

\item{lambda.S}{Desired scale parameter for the \code{Weibull} distribution associated with the Surrogate
endpoint. The default is 1.8.}

\item{nu.S}{Desired shape parameter for the \code{Weibull} distribution associated with the Surrogate
endpoint. The default is 0.0045.}

\item{lambda.T}{Desired scale parameter for the \code{Weibull} distribution associated with the True endpoint.
The default is 3.}

\item{nu.T}{Desired shape parameter for the \code{Weibull} distribution associated with the True endpoint.
The default is 0.0025.}

\item{ver}{Number of covariates. For surrogte evaluation, we just considered one covatiate, the treatment arm}

\item{typeOf}{Type of joint model used for data generation: 0 = classical joint model 
with a shared individual frailty effect (Rondeau, 2007), 1 = joint surrogate model with shared frailty 
effects \eqn{u_i} and \eqn{\omega_{ij}}, and two correlated random effects treatment-by-trial interaction 
\eqn{(v_{S_i},v_{T_i})} as described in Sofeu et al. (2018).}

\item{equi.subj.trial}{A binary variable that indicates if the same proportion of subjects should be included per trial (1) 
or not (0). If 0, the proportions of subject per trial are required in parameter \code{prop.subj.trial}.}

\item{equi.subj.trt}{A binary variable that indicates if the same proportion of subjects is randomized per trial (1) 
or not (0). If 0, the proportions of subject per trial are required in parameter \code{prop.subj.trt}.}

\item{prop.subj.trial}{The proportions of subjects per trial. Requires if \code{equi.subj.trial=0}.}

\item{prop.subj.trt}{The proportions of randomized subject per trial. Requires if \code{equi.subj.trt=0}.}

\item{full.data}{Specified if you wan the function to return the full dataset (1), including the random effects, 
or the restictive dataset (0) with \code{7} columns required for the function \code{\link{jointSurroPenal}}.}

\item{random.generator}{Random number generator to use by the Fortran compiler, 
\code{1} for the intrinsec subroutine \code{Random_number} and \code{2} for the 
subroutine \code{uniran()}. The default is \code{1}.}

\item{random}{A binary that says if we reset the random number generation with a different environment 
at each call \code{(1)} or not \code{(0)}. If it is set to \code{1}, we use the computer clock 
as seed. In the last case, it is not possible to reproduce the generated datasets". 
The default is \code{0}. Required if \code{random.generator} is set to 1.}

\item{random.nb.sim}{required if \code{random.generator} is set to 1, and if \code{random} is set to 1.}

\item{seed}{The seed to use for data (or samples) generation. Required if the argument \code{random.generator} is set to 1. 
Must be a positive value. If negative, the program do not account for seed. The default is \code{0}.}

\item{nb.reject.data}{Number of generation to reject before the considered dataset. this parameter is required
when data generation is for simulation. With a fixed parameter and \code{random.generator} set to 1,
all ganerated data are the same. By varying this parameter, different datasets are obtained during data genarations. The default value is 0, 
in case of one dataset.}
}
\value{
This function return if the parameter \code{full.data} is set to 0, a \code{\link{data.frame}} with columns :
   \item{patienID}{A numeric, that represents the patient's identifier, must be unique;}
   \item{trialID}{A numeric, that represents the trial in which each patient was randomized;}
   \item{trt}{The treatment indicator for each patient, with 1 = treated, 0 = untreated;}
   \item{timeS}{The follow up time associated with the surrogate endpoint;}
   \item{statusS}{The event indicator associated with the surrogate endpoint. Normally 
   0 = no event, 1 = event;}
   \item{timeT}{The follow up time associated with the true endpoint;}
   \item{statusT}{The event indicator associated with the true endpoint. Normally 
   0 = no event, 1 = event;}
If the argument \code{full.data} is set to 1, additionnal colums corresponding to random effects 
\eqn{\omega_{ij}}, \eqn{u_i}, \eqn{v_{S_i}} and \eqn{v_{T_i}} are returned. Note that
\eqn{u_i}, \eqn{v_{S_i}} and \eqn{v_{T_i}} are returned if \code{typeOf} is set to \code{1}
}
\description{
Date are generated from the one-step joint surrogate model (see \code{\link{jointSurroPenal}} for more details)
}
\details{
We just considered in this generation, the Gaussian random effects. If the parameter \code{full.data} is set to 1,
this function  return a list containning severals parameters, including the generated random effects. 
the desired individual level correlation (Kendall's \eqn{\tau}) depend on the values of 
\eqn{\alpha}, \eqn{\theta}, \eqn{\gamma} and \eqn{\zeta}.
}
\examples{

data.sim <- jointSurrSimul(n.obs=600, n.trial = 30,cens.adm=549.24, 
            alpha = 1.5, theta = 3.5, gamma = 2.5, sigma.s = 0.7, 
            zeta = 1, sigma.t = 0.7, rsqrt = 0.8, betas = -1.25, 
            betat = -1.25, full.data = 0, random.generator = 1, 
            seed = 0, nb.reject.data = 0)

}
\references{
Rondeau V., Mathoulin-Pelissier S., Jacqmin-Gadda H., Brouste V. and Soubeyran P. (2007).
Joint frailty models for recurring events and death using maximum penalized likelihood 
estimation: application on cancer events. Biostatistics 8(4), 708-721.

Sofeu C.L., Emura T. and Rondeau V. (2018). One-step validation method for surrogate 
endpoints in multiple randomized cancer clinical trials with failure-time endpoints. 
\code{Submitted}
}
\seealso{
\code{\link{jointSurrSimul}}
}
\author{
Casimir Ledoux Sofeu \email{casimir.sofeu@u-bordeaux.fr}, \email{scl.ledoux@gmail.com} and 
Virginie Rondeau \email{virginie.rondeau@inserm.fr}
}
