% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_val_croisee.r
\name{kappa_val_croisee}
\alias{kappa_val_croisee}
\title{Used to estimate smoothing parameters using cross-validation}
\usage{
kappa_val_croisee(don_S, don_T, njeu, n_obs, n_node = 6, adjust_S = 1,
  adjust_T = 1, kapp_0 = 100, kappa_1 = 955000, kappa_2 = 975000,
  print.times = T)
}
\arguments{
\item{don_S}{h}

\item{don_T}{h}

\item{njeu}{h}

\item{n_obs}{h}

\item{n_node}{h}

\item{adjust_S}{h}

\item{adjust_T}{h}

\item{kapp_0}{h}

\item{kappa_1}{h}

\item{kappa_2}{h}

\item{print.times}{h}
}
\value{
Create a text file named \bold{kappa_val_croisee.txt} in the work directory containing 
all estimated smoothing parameters. in addition, this function return an equivalent dataframe
of two coluumns for kappa associated with the surrogate endpoint and the true endpoint.
}
\description{
Function used to estimated smoothing parameters associated with the surrogate endpoint and the 
true endpoint by cross-validation. Are used two separate cox proportional hazard models.
This function created a file named \bold{kappa_val_croisee.txt} which contains the obtained
smoothing parameters. This is particularly useful when doing simulations study
}
