% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelp.R
\docType{data}
\name{longDat}
\alias{longDat}
\title{Longitudinal semicontinuous biomarker dataset (TPJM)}
\format{
This data frame contains the following columns: 
\describe{
\item{id}{The identification number of a patient}
\item{timej}{The measurement times of the biomarker}
\item{trtY}{Treatment covariate}
\item{Y}{Biomarker value}
}
}
\usage{
data(longDat)
}
\description{
This is a simulated dataset used to illustrate the two-part 
joint model included in the longiPenal function.
}
