% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{loocv}
\alias{loocv}
\title{Leave-one-out crossvalidation for the one-step Joint surrogate model for the evaluation of a 
canditate surrogate endpoint.}
\usage{
loocv(object, unusedtrial, var.used = "error.meta", alpha. = 0.05, print.times = TRUE)
}
\arguments{
\item{object}{An object inheriting from \code{jointSurroPenal} class
(output from calling \code{jointSurroPenal} function).}

\item{unusedtrial}{A list of trial not to be taken into account in the cross-validation}

\item{var.used}{This argument takes two values. The first one is \code{"error.meta"}
and indicates if the prediction error take into account
the estimation error of the estimates of the parameters. If the estimates 
are suppose knew or if the dataset includes a high number of trials with 
a high number of subject per trial, value \code{No.error} can be used. 
The default is \code{error.meta}.}

\item{alpha.}{The confidence level for the prediction interval. The default is \code{0.05}}

\item{print.times}{a logical parameter to print estimation time. Default is TRUE.}
}
\value{
Returns an object of class \code{jointSurroPenalloocv} containing a dataframe 
including for each trial the observed 
treatment effect on the surrogate endpoint, the observed treatment effect on
the true endpoint and the predicted treatment effect on the 
true enpoint with the associated prediction intervalls. If the observed treatment effect on the true 
endpoint is included into the prediction interval, the last columns contains "*".
}
\description{
{
Leave-one-out crossvalidation for the evaluation of the joint surrogate model 
}
}
\details{
{

}
}
\examples{


\dontrun{
# Generation of data to use 
 data.sim <- jointSurrSimul(n.obs=600, n.trial = 30,cens.adm=549.24, 
         alpha = 1.5, theta = 3.5, gamma = 2.5, zeta = 1, sigma.s = 0.7, 
         sigma.t = 0.7, rsqrt = 0.8, betas = -1.25, betat = -1.25, 
         full.data = 0, random.generator = 1, seed = 0, nb.reject.data = 0)

###--- Joint surrogate model ---###
 (Computation takes around 5 minutes)

joint.surro.sim.MCGH <- jointSurroPenal(data = data.sim, int.method = 2, 
                   nb.mc = 300, nb.gh = 20)
                
dloocv <- loocv(joint.surro.sim.MCGH, unusedtrial = 25)

}


}
\references{
Burzykowski T, Buyse M (2006). "Surrogate threshold effect: an alternative 
measure for meta-analytic surrogate endpoint validation." Pharmaceutical 
Statistics, 5(3), 173-186.ISSN 1539-1612.
}
\seealso{
\code{\link{jointSurroPenal}}
}
\author{
Casimir Ledoux Sofeu \email{casimir.sofeu@u-bordeaux.fr}, \email{scl.ledoux@gmail.com} and 
Virginie Rondeau \email{virginie.rondeau@inserm.fr}
}
\keyword{loocv}
\keyword{prediction}
\keyword{surrogate}
