% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.additivePenal.R
\name{print.additivePenal}
\alias{print.additivePenal}
\title{Print a Short Summary of parameter estimates of an additive frailty model}
\usage{
\method{print}{additivePenal}(x, digits = max(options()$digits - 4, 6),
...)
}
\arguments{
\item{x}{the result of a call to the additivePenal function}

\item{digits}{number of digits to print}

\item{\dots}{other unused arguments}
}
\value{
Print the parameter estimates of the survival or hazard functions.
}
\description{
Prints a short summary of the parameter estimates of an additive frailty
model or more generally of an 'additivePenal' object
}
\seealso{
\code{\link{additivePenal}}
}
\keyword{methods}
