% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.frailtyCmprsk.R
\name{print.frailtyCmprsk}
\alias{print.frailtyCmprsk}
\title{Print a Short Summary of parameter estimates of a Weibull competing risks model with (or without) shared frailty between transitions.}
\usage{
\method{print}{frailtyCmprsk}(x,
...)
}
\arguments{
\item{x}{the result of a call to the frailtyCmprsk function.}

\item{\dots}{other unused arguments.}
}
\value{
Print the parameter estimates of the survival or hazard functions.
}
\description{
Prints a short summary of parameter estimates of a 'frailtyCmprsk' object
}
\seealso{
\code{\link{frailtyCmprsk}}
}
\keyword{methods}
