% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.frailtyIllnessDeath.R
\name{print.frailtyIllnessDeath}
\alias{print.frailtyIllnessDeath}
\title{Print a Short Summary of parameter estimates of a Weibull Illness-Death model with (or without) shared frailty between transitions.}
\usage{
\method{print}{frailtyIllnessDeath}(x,
...)
}
\arguments{
\item{x}{the result of a call to the frailtyIllnessDeath function.}

\item{\dots}{other unused arguments.}
}
\value{
Print the parameter estimates of the survival or hazard functions.
}
\description{
Prints a short summary of parameter estimates of a 'frailtyIllnessDeath' object
}
\seealso{
\code{\link{frailtyIllnessDeath}}
}
\keyword{methods}
