% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.frailtyPenal.R
\name{print.frailtyPenal}
\alias{print.frailtyPenal}
\title{Print a Short Summary of parameter estimates of a shared frailty model}
\usage{
\method{print}{frailtyPenal}(x, digits = max(options()$digits - 4, 6),
...)
}
\arguments{
\item{x}{the result of a call to the frailtyPenal function.}

\item{digits}{number of digits to print.}

\item{\dots}{other unused arguments.}
}
\value{
Print the parameter estimates of the survival or hazard functions.
}
\description{
Prints a short summary of parameter estimates of a 'frailtyPenal' object
}
\seealso{
\code{\link{frailtyPenal}}
}
\keyword{methods}
