% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.longiPenal.R
\name{print.longiPenal}
\alias{print.longiPenal}
\title{Print a Summary of parameter estimates of a joint model for longitudinal
data and a terminal event}
\usage{
\method{print}{longiPenal}(x, digits = max(options()$digits - 4, 6), ...)
}
\arguments{
\item{x}{an object inheriting from \code{longiPenal} class}

\item{digits}{number of digits to print}

\item{\dots}{other unused arguments}
}
\value{
Print, separately for each part of the model (longitudinal and terminal) the
parameter estimates and details on the estimation.
}
\description{
Prints a short summary of parameter estimates of a joint model for
longitudinal data and a terminal event, an object inheriting from class
'longiPenal'.
}
\seealso{
\code{\link{longiPenal}}
}
\keyword{methods}
