% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trivPenal.R
\name{print.trivPenal}
\alias{print.trivPenal}
\title{Print a Summary of parameter estimates of a joint model for longitudinal
data, recurrent events and a terminal event}
\usage{
\method{print}{trivPenal}(x, digits = max(options()$digits - 4, 6), ...)
}
\arguments{
\item{x}{an object inheriting from \code{trivPenal} class}

\item{digits}{number of digits to print}

\item{\dots}{other unused arguments}
}
\value{
Print, separately for each part of the model (longitudinal, recurrent and
terminal) the parameter estimates and details on the estimation.
}
\description{
Prints a short summary of parameter estimates of a joint model for
longitudinal data, recurrent events and a terminal event, an object
inheriting from class 'trivPenal'.
}
\seealso{
\code{\link{trivPenal}}
}
\keyword{methods}
