% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Shiny application for modelisation and prediction of frailty models}
\usage{
runShiny()
}
\value{
No value returned.
}
\description{
This function loads the shiny package and runs the application for modelisation and prediction of several frailty models using package frailtypack.
}
\examples{
\dontrun{

runShiny()

}

}
\references{
Rizopoulos D. (2016)
}
