% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jointRecCompet.R
\name{summary.jointRecCompet}
\alias{summary.jointRecCompet}
\title{Summary method for a joint competing risks midel}
\usage{
\method{summary}{jointRecCompet}(object, digits = max(options()$digits - 4, 6),
...)
}
\arguments{
\item{object}{the result of a call to the jointRecCompet function}

\item{digits}{number of digits to print}

\item{\dots}{other unused arguments}
}
\value{
Print, separately for each type of event (Recurrent, Terminal1 and
Terminal2), the parameter estimates of the survival or hazard functions.
}
\description{
Prints a short summary of parameter estimates of a joint competing risks model
or more generally an object of class 'jointRecCompet'.
}
\seealso{
\code{\link{jointRecCompet}}
}
\keyword{jointRecCompet}
\keyword{methods}
