% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHelp.R
\docType{data}
\name{survDat}
\alias{survDat}
\title{Survival dataset (TPJM)}
\format{
This data frame contains the following columns: 
\describe{
\item{id}{The identification number of a patient}
\item{deathTimes}{The event times (death or censoring)}
\item{d}{Censoring indicator}
\item{trt}{Treatment covariate}
}
}
\usage{
data(survDat)
}
\description{
This is a simulated dataset used to illustrate the two-part 
joint model included in the longiPenal function.
}
