% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testOpenMPFortran.R
\name{testOpenMPFortran}
\alias{testOpenMPFortran}
\title{Used to teste pactice use of OpenMP with Fortran code in an R package.}
\usage{
testOpenMPFortran(a,b,nboot,nbThread,stat=T)
}
\arguments{
\item{a}{The first term of the sum}

\item{b}{The second term of the sum}

\item{nboot}{The maximum term of the sequence}

\item{nbThread}{The number of thread to used in the OpenMP procedure}

\item{stat}{A logical specifying if you want to print the result of the sum on screen or not}
}
\description{
In this function we test a call of a fortran subroutine in a R package. we therefore compute a sum of two numbers 
using a fortran subroutine. In addition, we test the use of OpenMP by compute an arithmetic Sequences with first term 1 
and the commun difference 1.
}
\examples{

\dontrun{
frailtypack:::testOpenMPFortran(2,4,10000,10,stat=T)

}
}
\seealso{
\code{\link{evalOpenMPFortran}}
}
\author{
Casimir Ledoux SOFEU <scl.ledoux@gmail.com>
}
\keyword{Fortran}
\keyword{OpenMP}
\keyword{R}
