% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event2.R
\name{event2}
\alias{event2}
\title{Identify event2 indicator}
\usage{
event2(x)
}
\arguments{
\item{x}{A numeric variable but should be a boolean which equals 1 if the
subject has experienced an event of type 2 and 0 if not.}
}
\value{
\item{x}{an indicator for an event of type 2}
}
\description{
This is a special function used in the context of multivariate frailty model
with two types of recurrent events and a terminal event (e.g., censoring
variable related to both recurrent events). It contains the indicator of the
recurrent event of type 2, normally 0=no event, 1=event, and is used on the
right hand side of a formula of a 'multivPenal' object.  Using
\code{event2()} in a formula implies that a multivariate frailty model for
two types of recurrent events and a terminal event is fitted.
}
\seealso{
\code{\link{multivPenal}}
}
\keyword{misc}
