\name{print.Cmeasures}
\alias{print.Cmeasures}

\title{
Print a short summary of results of Cmeasure function.
}
\description{
Print a short summary of results of the concordance measure estimated by the Cmeasure function.
}
\usage{
\method{print}{Cmeasures}(x, ...)
}

\arguments{
  \item{x}{a Cmeasures object.}
  \item{\dots}{Other unused arguments}
}

\value{

Print concordance measures estimated.

}

\seealso{
 \code{\link{Cmeasures}}
}

\keyword{methods}

