% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Cmeasures.R
\name{print.Cmeasures}
\alias{print.Cmeasures}
\title{Print a short summary of results of Cmeasure function.}
\usage{
\method{print}{Cmeasures}(x, ...)
}
\arguments{
\item{x}{a Cmeasures object.}

\item{\dots}{Other unused arguments}
}
\value{
Print concordance measures estimated.
}
\description{
Print a short summary of results of the concordance measure estimated by the
Cmeasure function.
}
\seealso{
\code{\link{Cmeasures}}
}
\keyword{methods}
