% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.multivPenal.R
\name{summary.multivPenal}
\alias{summary.multivPenal}
\title{summary of parameter estimates of a multivariate frailty model.}
\usage{
\method{summary}{multivPenal}(object, level = 0.95, len = 6, d = 2, lab
= "hr", ...)
}
\arguments{
\item{object}{output from a call to multivPenal for joint multivariate
models}

\item{level}{significance level of confidence interval. Default is 95\%.}

\item{len}{the total field width. Default is 6.}

\item{d}{the desired number of digits after the decimal point. Default of 6
digits is used.}

\item{lab}{label of printed results.}

\item{\dots}{other unused arguments.}
}
\value{
Prints HR and its confidence intervals for each covariate.
Confidence level is allowed (level argument)
}
\description{
This function returns hazard ratio (HR) and its confidence intervals.
}
\seealso{
\code{\link{multivPenal}}
}
\keyword{methods}
\keyword{multiv}
